+++ hookee/manager.py
@@ -13,7 +13,7 @@ from pluginbase import PluginBase
 
 __author__ = "Alex Laird"
 __copyright__ = "Copyright 2020, Alex Laird"
-__version__ = "0.0.4"
+__version__ = "0.0.7"
 
 
 class Manager:
@@ -21,6 +21,8 @@ class Manager:
         self.ctx = ctx
 
         self.config = conf.Config(self.ctx)
+        self.last_request = None
+        self.last_response = None
         self.loaded_plugins = self.load_plugins()
         self.print_util = PrintUtil(self.config)
 
@@ -35,24 +37,28 @@ class Manager:
 
     def start(self):
         if not self.alive:
-            self.server.start()
-            self.tunnel.start()
-            self.alive = True
-
-            self.print_ready()
-
             try:
+                self.alive = True
+                self.server.start()
+                self.tunnel.start()
+
+                self.print_ready()
+
                 while self.alive:
                     time.sleep(1)
             except KeyboardInterrupt:
                 pass
+                # self.alive = False
+                # self.server.stop()
+                # self.tunnel.stop()
 
             self.stop()
 
     def stop(self):
         if self.alive:
             self.server.stop()
-            self.tunnel._thread.alive = False
+            if self.tunnel._thread:
+                self.tunnel._thread.alive = False
 
             # Wait for the other threads to teardown
             while self.server._thread and self.tunnel._thread:
@@ -75,28 +81,26 @@ class Manager:
         click.secho("--> Ready, send a request to a registered endpoint ...", fg="green", bold=True)
         click.echo("")
 
-    def validate_plugin(self, plugin_name):
+    def validate_plugin(self, plugin):
         try:
-            plugin = self.source.load_plugin(plugin_name)
-
             functions_list = [o[0] for o in inspect.getmembers(plugin, inspect.isfunction)]
             attributes = dir(plugin)
 
             if not all(elem in attributes for elem in ["plugin_type", "manager"]):
-                self.ctx.fail("Plugin \"{}\" does not conform to the plugin spec.".format(plugin_name))
+                self.ctx.fail("Plugin \"{}\" does not conform to the plugin spec.".format(plugin.__name__))
 
             if plugin.plugin_type in [util.REQUEST_PLUGIN, util.RESPONSE_PLUGIN]:
                 all(elem in functions_list for elem in ["setup", "run"])
             elif plugin.plugin_type == util.BLUEPRINT_PLUGIN:
                 all(elem in functions_list for elem in ["setup"])
             else:
-                self.ctx.fail("Plugin \"{}\" is not a valid plugin type.".format(plugin_name))
+                self.ctx.fail("Plugin \"{}\" is not a valid plugin type.".format(plugin.__name__))
 
             # TODO: additionally validate the functions have correct num args
 
             return plugin
         except ModuleNotFoundError:
-            self.ctx.fail("Plugin \"{}\" could not be found.".format(plugin_name))
+            self.ctx.fail("Plugin \"{}\" could not be found.".format(plugin.__name__))
 
     def load_plugins(self):
         builtin_plugins_dir = os.path.normpath(os.path.join(os.path.abspath(os.path.dirname(__file__)), "plugins"))
@@ -109,9 +113,18 @@ class Manager:
 
         loaded_plugins = []
         for plugin_name in enabled_plugins:
-            plugin = self.validate_plugin(plugin_name)
+            plugin = self.validate_plugin(self.source.load_plugin(plugin_name))
             plugin.setup(self)
             loaded_plugins.append(plugin)
+
+        last_request = self.config.get("last_request")
+        if last_request:
+            self.last_request = __import__(last_request.__name__, globals(), {}, ['__name__'])
+            self.validate_plugin(self.last_request)
+        last_response = self.config.get("last_response")
+        if last_response:
+            self.last_response = __import__(last_response.__name__, globals(), {}, ['__name__'])
+            self.validate_plugin(self.last_response)
 
         return loaded_plugins